+++ crates/api/src/callable.rs
@@ -1,12 +1,5 @@
-use crate::runtime::Store;
-use crate::trampoline::generate_func_export;
 use crate::trap::Trap;
-use crate::types::FuncType;
 use crate::values::Val;
-use std::cmp::max;
-use std::ptr;
-use std::rc::Rc;
-use wasmtime_runtime::{ExportFunction, InstanceHandle, VMTrampoline};
 
 /// A trait representing a function that can be imported and called from inside
 /// WebAssembly.
@@ -83,132 +76,4 @@ pub trait Callable {
     /// `results` is mutable list of results to be potentially set by your
     /// function. Produces a `Trap` if the function encounters any errors.
     fn call(&self, params: &[Val], results: &mut [Val]) -> Result<(), Trap>;
-}
-
-pub(crate) trait WrappedCallable {
-    fn call(&self, params: &[Val], results: &mut [Val]) -> Result<(), Trap>;
-    fn wasmtime_handle(&self) -> &InstanceHandle;
-    fn wasmtime_function(&self) -> &ExportFunction;
-}
-
-pub(crate) struct WasmtimeFn {
-    store: Store,
-    instance: InstanceHandle,
-    export: ExportFunction,
-    trampoline: VMTrampoline,
-}
-
-impl WasmtimeFn {
-    pub fn new(
-        store: &Store,
-        instance: InstanceHandle,
-        export: ExportFunction,
-        trampoline: VMTrampoline,
-    ) -> WasmtimeFn {
-        WasmtimeFn {
-            store: store.clone(),
-            instance,
-            export,
-            trampoline,
-        }
-    }
-}
-
-impl WrappedCallable for WasmtimeFn {
-    fn call(&self, params: &[Val], results: &mut [Val]) -> Result<(), Trap> {
-        let f = self.wasmtime_function();
-        let signature = self
-            .store
-            .compiler()
-            .signatures()
-            .lookup(f.signature)
-            .expect("missing signature");
-        if signature.params.len() - 2 != params.len() {
-            return Err(Trap::new(format!(
-                "expected {} arguments, got {}",
-                signature.params.len() - 2,
-                params.len()
-            )));
-        }
-        if signature.returns.len() != results.len() {
-            return Err(Trap::new(format!(
-                "expected {} results, got {}",
-                signature.returns.len(),
-                results.len()
-            )));
-        }
-
-        let mut values_vec = vec![0; max(params.len(), results.len())];
-
-        // Store the argument values into `values_vec`.
-        let param_tys = signature.params.iter().skip(2);
-        for ((arg, slot), ty) in params.iter().zip(&mut values_vec).zip(param_tys) {
-            if arg.ty().get_wasmtime_type() != Some(ty.value_type) {
-                return Err(Trap::new("argument type mismatch"));
-            }
-            unsafe {
-                arg.write_value_to(slot);
-            }
-        }
-
-        // Call the trampoline.
-        if let Err(error) = unsafe {
-            wasmtime_runtime::wasmtime_call_trampoline(
-                f.vmctx,
-                ptr::null_mut(),
-                self.trampoline,
-                f.address,
-                values_vec.as_mut_ptr() as *mut u8,
-            )
-        } {
-            return Err(Trap::from_jit(error));
-        }
-
-        // Load the return values out of `values_vec`.
-        for (index, abi_param) in signature.returns.iter().enumerate() {
-            unsafe {
-                let ptr = values_vec.as_ptr().add(index);
-
-                results[index] = Val::read_value_from(ptr, abi_param.value_type);
-            }
-        }
-
-        Ok(())
-    }
-    fn wasmtime_handle(&self) -> &InstanceHandle {
-        &self.instance
-    }
-    fn wasmtime_function(&self) -> &ExportFunction {
-        &self.export
-    }
-}
-
-pub struct NativeCallable {
-    callable: Rc<dyn Callable + 'static>,
-    instance: InstanceHandle,
-    export: ExportFunction,
-}
-
-impl NativeCallable {
-    pub(crate) fn new(callable: Rc<dyn Callable + 'static>, ft: &FuncType, store: &Store) -> Self {
-        let (instance, export) =
-            generate_func_export(ft, &callable, store).expect("generated func");
-        NativeCallable {
-            callable,
-            instance,
-            export,
-        }
-    }
-}
-
-impl WrappedCallable for NativeCallable {
-    fn call(&self, params: &[Val], results: &mut [Val]) -> Result<(), Trap> {
-        self.callable.call(params, results)
-    }
-    fn wasmtime_handle(&self) -> &InstanceHandle {
-        &self.instance
-    }
-    fn wasmtime_function(&self) -> &ExportFunction {
-        &self.export
-    }
 }
